+++ internal/helper/helper.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 
 	"github.com/go-enry/go-license-detector/v4/licensedb"
-	"github.com/opensbom-generator/parsers/license"
+	"github.com/opensbom-generator/parsers/internal/license"
 
 	log "github.com/sirupsen/logrus"
 )
